refactor(dfpt): extract DFPT_PW::solve_displacement + assemble_v_sc helpers
Moves DFPT_PW::Impl::solve_displacement from dfpt_pw.cpp into
dfpt_pw_solve.cpp together with the shared Impl::assemble_v_sc
definition (declared in dfpt_pw_impl.h so the efield solve in the
next split can reuse it without code duplication). The 36-path
solve_displacement body is decomposed into:
  * assemble_v_sc (public Impl helper, already declared in impl.h)
  * sternheimer_per_band  -- private per-(ik,ib) dispatch
  * stash_converged_disp_response -- private stash helper
  * debug_iter_snapshot / debug_h_consistency (free helpers in an
    anonymous namespace to keep the main flow linear)
Every moved function stays under the cyclomatic complexity 10 budget
and the new translation unit is 266 lines.

Verification commands:
  cmake --build build -j 30                         -> PASS
  cmake --build build_max_para_test -j 30 \
    --target MODULE_DFPT_pw_run_test                -> PASS
  cmake --build build_max_para_test -j 30 \
    --target MODULE_DFPT_{pw_data_test,pert_serial,\
    phon_serial,q0_serial,rho_serial}               -> PASS all
  OMP_NUM_THREADS=1 ctest --test-dir build_max_para_test \
    -R MODULE_DFPT                                  -> 8/8 Passed (5.24 s)
  python3 tools/03_code_analysis/agent_governance_check.py \
    --staged                                        -> 0 deterministic
                                                       errors (2 PR-body
                                                       warnings only:
                                                       evidence review,
                                                       docs sync; refactor
                                                       with 100% API
                                                       compatible, no
                                                       behavior change)
