+++ source/source_pw/module_dfpt/dfpt_pw.cpp
@@ -61,220 +61,6 @@ bool DFPT_PW::get_u_active() const
     return pimpl_->data_.u_active();
 }
 
-double DFPT_PW::Impl::solve_displacement(int q_idx, int iat, int idir)
-{
-    ModuleBase::TITLE("DFPT_PW", "solve_displacement");
-    ModuleBase::timer::start("DFPT_PW", "solve_displacement");
-    if (!wired() || hamilt_ == nullptr)
-    {
-        ModuleBase::timer::end("DFPT_PW", "solve_displacement");
-        return 0.0;
-    }
-    const ModuleBase::Vector3<double> q_frac = data_.get_qvec(q_idx);
-    const ModuleBase::Vector3<double> q_cart = q_frac * ucell_->G;
-    const int nrxx = pw_rho_->nrxx;
-    const int nk = gs_psi_.get_nk();
-    const int nbands = gs_psi_.get_nbands();
-
-    pert_.build_dv(q_idx, iat, idir, data_);
-    rho_.reset_mixing(q_idx);
-    // the previous perturbation's stored response must not leak into the
-    // first iteration of this one
-    data_.set_drho_g(q_idx, 0, std::vector<std::complex<double>>(pw_rho_->npw, std::complex<double>(0.0, 0.0)));
-
-    const int lin_max = data_.get_max_iter();
-    const double lin_thr = data_.get_conv_thr();
-
-    bool converged = false;
-    double residual = 0.0;
-    const bool dbg = (getenv("DFPT_DEBUG") != nullptr);
-    // last screened response potential (hoisted out of the loop: the 2n+1
-    // accumulation below needs the converged v_sc of this displacement)
-    std::vector<std::complex<double>> v_sc_r_last;
-    for (int iter = 0; iter < max_iter_ && !converged; ++iter)
-    {
-        // the per-displacement SCF state (iter / residual / converged) is
-        // local to this solve: the DFPT_PW_Data ledger is the per-(q,irrep)
-        // outer-pass record kept by run(), and the final residual is
-        // returned to the caller for that aggregation (B4)
-
-        // ---- 1. screened response potential from the mixed input density:
-        // q-shifted complex periodic amplitude on the shared grid, i.e. the
-        // same convention as dv_rc (v_hartree_q acts on the q-shifted
-        // coefficients; the XC kernel responds to Re/Im of the amplitude)
-        std::vector<std::complex<double>> v_sc_r(nrxx, std::complex<double>(0.0, 0.0));
-        const std::vector<std::complex<double>> drho_in_g = data_.get_drho_g(q_idx, 0);
-        if (!drho_in_g.empty() && static_cast<int>(drho_in_g.size()) == pw_rho_->npw)
-        {
-            std::vector<std::complex<double>> dv_ha_g;
-            rho_.v_hartree_q(q_cart, drho_in_g, dv_ha_g);
-            std::vector<std::complex<double>> vh_r(nrxx);
-            pw_rho_->recip2real(dv_ha_g.data(), vh_r.data());
-            for (int ir = 0; ir < nrxx; ++ir)
-            {
-                v_sc_r[ir] = vh_r[ir];
-            }
-            if (xc_ != nullptr)
-            {
-                std::vector<std::complex<double>> a_r(nrxx);
-                pw_rho_->recip2real(drho_in_g.data(), a_r.data());
-                std::vector<std::complex<double>> b_r;
-                xc_->apply(a_r, b_r);
-                if (static_cast<int>(b_r.size()) == nrxx)
-                {
-                    for (int ir = 0; ir < nrxx; ++ir)
-                    {
-                        v_sc_r[ir] += b_r[ir];
-                    }
-                }
-            }
-            if (dbg)
-            {
-                double dh = 0.0;
-                double dv = 0.0;
-                for (int ig = 0; ig < pw_rho_->npw; ++ig)
-                {
-                    dh += std::norm(drho_in_g[ig]);
-                }
-                for (int ir = 0; ir < nrxx; ++ir)
-                {
-                    dv += std::norm(v_sc_r[ir]);
-                }
-                std::cout << "DBG iter=" << iter << " |drho_in_g|=" << std::sqrt(dh) << " |v_sc_r|=" << std::sqrt(dv)
-                          << std::endl;
-            }
-        }
-        v_sc_r_last = v_sc_r;
-
-        // ---- 2. Sternheimer solve of every occupied (k, band)
-        for (int ik = 0; ik < nk; ++ik)
-        {
-            if (static_cast<int>(occ_kq_.size()) <= ik || occ_kq_[ik].empty())
-            {
-                if (dbg)
-                {
-                    std::cout << "DBG skip ik=" << ik << " no occ_kq" << std::endl;
-                }
-                continue; // no occupied states at k+q: nothing to solve
-            }
-            // dV_ext |psi_n> for all bands (dVloc convolution + dVnl_dtau)
-            pert_.apply_dv(q_idx, ik, gs_psi_, data_);
-            // screened response part |v_sc psi_n>
-            std::vector<std::vector<std::complex<double>>> dv_sc;
-            pert_.apply_vr(q_idx, ik, v_sc_r, gs_psi_, q_cart, dv_sc);
-            if (ik != last_ik_ || last_q_ != q_idx)
-            {
-                hamilt_->set_context(q_cart, ik);
-                last_ik_ = ik;
-                if (dbg)
-                {
-                    std::cout << "DBG occ_kq nstates=" << occ_kq_[ik].size() << std::endl;
-                    for (size_t m = 0; m < occ_kq_[ik].size(); ++m)
-                    {
-                        double nrm = 0.0;
-                        for (size_t i = 0; i < occ_kq_[ik][m].size(); ++i)
-                        {
-                            nrm += std::norm(occ_kq_[ik][m][i]);
-                        }
-                        std::cout << "DBG   occ[" << m << "] |psi|^2=" << nrm << std::endl;
-                    }
-                    // kernel consistency: <psi_m|H(k+q)|psi_m> must equal
-                    // eig(ikq, m); the eigenvalue used by set_shift below is
-                    // the k-side one (equal only when H is assembled right)
-                    for (size_t m = 0; m < occ_kq_[ik].size(); ++m)
-                    {
-                        hamilt_->set_shift(0.0);
-                        std::vector<std::complex<double>> hp(occ_kq_[ik][m].size());
-                        hamilt_->apply(occ_kq_[ik][m].data(), hp.data());
-                        std::complex<double> dot(0.0, 0.0);
-                        for (size_t i = 0; i < hp.size(); ++i)
-                        {
-                            dot += std::conj(occ_kq_[ik][m][i]) * hp[i];
-                        }
-                        std::cout << "DBG   <psi_" << m << "|H|psi_" << m << "> = " << dot.real() << " + i "
-                                  << dot.imag() << "  (GS eig " << eig_(ikq_of_k_[ik], static_cast<int>(m)) << ")"
-                                  << std::endl;
-                        std::cout << "DBG   <psi_" << m << "|T+Vnl|psi_" << m
-                                  << "> = " << hamilt_->debug_t_vnl(occ_kq_[ik][m]) << std::endl;
-                        std::cout << "DBG   <psi_" << m << "|V(wfc-path)|psi_" << m
-                                  << "> = " << hamilt_->debug_v_wfc(occ_kq_[ik][m]) << std::endl;
-                    }
-                }
-            }
-            for (int ib = 0; ib < nbands; ++ib)
-            {
-                if (!dfpt_band_occupied(wg_, ik, ib))
-                {
-                    continue; // unoccupied: no Sternheimer equation
-                }
-                std::vector<std::complex<double>> rhs = data_.get_dpsi(q_idx, ik, ib);
-                if (rhs.empty() || static_cast<int>(dv_sc.size()) != nbands || rhs.size() != dv_sc[ib].size())
-                {
-                    if (dbg)
-                    {
-                        std::cout << "DBG skip solve ik=" << ik << " ib=" << ib << " rhs.size=" << rhs.size()
-                                  << " dv_sc.size=" << dv_sc.size() << " dv_sc[ib].size="
-                                  << (dv_sc.size() > static_cast<size_t>(ib) ? dv_sc[ib].size() : 999999) << std::endl;
-                    }
-                    continue;
-                }
-                // b = -(dV_ext + dV_sc)|psi_n>
-                for (size_t i = 0; i < rhs.size(); ++i)
-                {
-                    rhs[i] = -(rhs[i] + dv_sc[ib][i]);
-                }
-                hamilt_->set_shift(eig_(ik, ib));
-                std::vector<std::complex<double>> dpsi_out;
-                double res = 0.0;
-                stern_.solve(*hamilt_, occ_kq_[ik], rhs, lin_max, lin_thr, dpsi_out, res);
-                if (dbg)
-                {
-                    double nr = 0.0, nb2 = 0.0;
-                    for (size_t i = 0; i < dpsi_out.size(); ++i)
-                    {
-                        nr += std::norm(dpsi_out[i]);
-                        nb2 += std::norm(rhs[i]);
-                    }
-                    std::cout << "DBG solve ik=" << ik << " ib=" << ib << " eps=" << eig_(ik, ib) << " res=" << res
-                              << " |dpsi|=" << std::sqrt(nr) << " |rhs|=" << std::sqrt(nb2)
-                              << " finite=" << (std::isfinite(std::sqrt(nr)) ? 1 : 0) << std::endl;
-                }
-                data_.set_dpsi(q_idx, ik, ib, dpsi_out);
-            }
-        }
-
-        // ---- 3. first-order density and mixing
-        rho_.compute_drho(gs_psi_, wg_, q_idx, data_);
-        rho_.mix_drho(q_idx, data_);
-        residual = rho_.get_residual(q_idx, data_);
-        if (dbg)
-        {
-            std::cout << "DBG iter=" << iter << " residual=" << residual << " conv_thr=" << conv_thr_ << std::endl;
-        }
-        converged = (residual < conv_thr_);
-    }
-    // stash the converged screened potential and dpsi of this displacement
-    // for the two-pass 2n+1 accumulation (term2 cross section needs
-    // dV_ext^b + dV_sc^b and dpsi^b of every displacement)
-    data_.set_vsc_r(iat, idir, v_sc_r_last);
-    {
-        std::vector<std::vector<std::vector<std::complex<double>>>> disp(
-            nk,
-            std::vector<std::vector<std::complex<double>>>(nbands));
-        for (int ik = 0; ik < nk; ++ik)
-        {
-            for (int ib = 0; ib < nbands; ++ib)
-            {
-                disp[ik][ib] = data_.get_dpsi(q_idx, ik, ib);
-            }
-        }
-        data_.set_dpsi_disp(iat, idir, disp);
-    }
-
-    ModuleBase::timer::end("DFPT_PW", "solve_displacement");
-    return residual;
-}
-
 void DFPT_PW::Impl::solve_pos_resp(int q_idx)
 {
     ModuleBase::TITLE("DFPT_PW", "solve_pos_resp");
