+++ source/source_pw/module_dfpt/test/CMakeLists.txt
@@ -40,6 +40,7 @@ AddTest(
 	../dfpt_pw.cpp
 	../dfpt_pw_init.cpp
 	../dfpt_pw_run.cpp
+	../dfpt_pw_solve.cpp
 	../dfpt_pw_data.cpp
 	../dfpt_pert.cpp
 	../dfpt_pert_vkb.cpp
